javadoc update

+++ neuroph-2.9/Core/src/main/java/org/neuroph/core/learning/error/MeanSquaredError.java
@@ -32,7 +32,7 @@ public final class MeanSquaredError implements ErrorFunction, Serializable {
     /**
      * Number of patterns - n 
      */
-    private transient double patternCount;
+    private transient float patternCount;
 
     public MeanSquaredError() {
         reset();
